[SYCL][FPGA][NFC] Refactor [[intel::max_global_work_dim()]] attribute tests
This patch updates tests for [[intel::max_global_work_dim()]] attribute:

1. separates AST and diagnostics to make it easier to read/follow-up.
2. removes duplicate test cases.
3. adds missing tests.
4. no compiler changes.
